refactor(proxy): extract row codec helpers
Move MySQL record and LineairDB row-byte conversion helpers into row_codec.cc.
Register row_codec.cc in the proxy build.
