+++ proxy/CMakeLists.txt
@@ -43,6 +43,7 @@ set(LINEAIRDB_SOURCES ha_lineairdb.cc ha_lineairdb.hh
                       lineairdb_transaction.cc lineairdb_transaction.hh
                       lineairdb_proxy.cc lineairdb_proxy.hh
                       lineairdb_keyenc.cc lineairdb_keyenc.hh
+                      row_codec.cc
                       lineairdb_pushdown.cc lineairdb_pushdown.hh
                       lineairdb_index_search.cc lineairdb_index_search.hh
                       lineairdb_prefetch.cc lineairdb_prefetch.hh
